refactor: 기존 코드에 맞게 ProductIntegrationTest의 테스트 코드 수정
