+++ src/test/java/com/gugucon/shopping/utils/DomainUtils.java
@@ -36,6 +36,28 @@ public class DomainUtils {
                 .build();
     }
 
+    public static Product createProductWithoutId(String name) {
+        sequence++;
+        return Product.builder()
+                .name(name)
+                .imageFileName("image_file_name_" + sequence)
+                .stock(100)
+                .description("test_description")
+                .price(1000L)
+                .build();
+    }
+
+    public static Product createProductWithoutId(String name, long price) {
+        sequence++;
+        return Product.builder()
+                .name(name)
+                .imageFileName("image_file_name_" + sequence)
+                .stock(100)
+                .description("test_description")
+                .price(price)
+                .build();
+    }
+
     public static Product createSoldOutProduct(String name, long price) {
         sequence++;
         return Product.builder()
